+++ app/models/psql.js
@@ -41,6 +41,41 @@ types.setTypeParser(1022, floatArrayParser); // _float8
 types.setTypeParser(1231, floatArrayParser); // _numeric
 types.setTypeParser(1016, floatArrayParser); // _int8
 
+// Standard type->name mappnig (up to oid=2000)
+var stdTypeName = {
+    16: 'bool',
+    17: 'bytea',
+    20: 'int8',
+    21: 'int2',
+    23: 'int4',
+    25: 'text',
+    26: 'oid',
+   114: 'JSON',
+   700: 'float4',
+   701: 'float8',
+  1000: '_bool',
+  1015: '_varchar',
+  1042: 'bpchar',
+  1043: 'varchar',
+  1005: '_int2',
+  1007: '_int4',
+  1014: '_bpchar',
+  1016: '_int8',
+  1021: '_float4',
+  1022: '_float8',
+  1008: '_regproc',
+  1009: '_text',
+  1114: 'timestamp',
+  1182: 'date',
+  1184: 'timestampz',
+  1186: 'interval',
+  1231: '_numeric',
+  1700: 'numeric'
+};
+
+// Holds a typeId->typeName mapping for each
+// database ever connected to
+var extTypeName = {};
 
 // PSQL
 //
@@ -80,6 +115,41 @@ var PSQL = function(user_id, db) {
                     global.settings.db_port + "/" +
                     me.database();
 
+    me.ensureTypeCache = function(cb) {
+      var db = this.db;
+      if ( extTypeName[db] ) { cb(); return; }
+      pg.connect(this.conString, function(err, client, done) {
+        if ( err ) { cb(err); return; }
+        var types = ["'geometry'","'raster'"]; // types of interest
+        client.query("SELECT oid, typname FROM pg_type where typname in (" + types.join(',') + ")", function(err,res) {
+          done();
+          if ( err ) { cb(err); return; }
+          var cache = {};
+          res.rows.map(function(r) {
+            cache[r.oid] = r.typname;
+          });
+          extTypeName[db] = cache;
+          cb();
+        });
+      });
+    }
+
+    // Return type name for a type identifier
+    //
+    // Possibly returns undefined, for unkonwn (uncached)
+    //
+    me.typeName = function(typeId) {
+      return stdTypeName[typeId] ? stdTypeName[typeId] : extTypeName[this.db][typeId];
+    }
+
+    me.connect = function(cb){
+      var that = this;
+      this.ensureTypeCache(function(err) {
+        if ( err ) cb(err);
+        else pg.connect(that.conString, cb);
+      });
+    };
+
     me.eventedQuery = function(sql, callback){
         var that = this;
 
@@ -89,7 +159,7 @@ var PSQL = function(user_id, db) {
             },
             function(err, clean){
                 if (err) throw err;
-                pg.connect(that.conString, this);
+                that.connect(this);
             },
             function(err, client, done){
                 if (err) throw err;
@@ -116,7 +186,7 @@ var PSQL = function(user_id, db) {
             },
             function(err, clean){
                 if (err) throw err;
-                pg.connect(that.conString, this);
+                that.connect(this);
             },
             function(err, client, done){
                 if (err) throw err;